Fixed bug comments
Sometimes, the comment section was empty when the current taxon was changed. The reason was because the model of the html document was changed twice. Once when, all comments are handled in setComments() method and once just after the setComments method was called.
This problem allowed us to reactivate the caching of comments in memory.