Fixed bug with graph and add new control.
The problem was the original graph displayed in the form was not a graph control but a graphPanel control. Consequently, the selection of a taxon in the graph didn't trigger the change of the currentTaxon property in TaxonUtils.
A button to add a new tree was also added in the File menu.