Reset NewTaxon form instead of AddTaxon
In commit '14de33' the form opened when clicking 'Add child' was changed from NewTaxon to AddTaxon. However, after discussions with project supervision team, this new form complicated the process of creation of a tree too much. So we restore the first one.
We also changed the form opened when clicking 'file -> new' in the main menu to this form.